scheduling update
new properties inside of versions to track the date the goal was created, the schedule date for when it should completed, and a variable (resomethingstupid)  to reset the schedule based off interval.

if the schedule is set to null it is exected that the idea does not have time restraints, and therefore also does not have the re variable.